BB-547 : Download catchment project list on excel

+++ data/src/main/java/com/proptiger/data/repo/trend/TrendReportDao.java
@@ -310,7 +310,7 @@ public class TrendReportDao {
             }
             else {
                 Collections.sort(bhkSizeList);
-                return (bhkSizeList.get(0) + "-" + bhkSizeList.get(bhkSizeList.size() - 1));
+                return (bhkSizeList.get(0).intValue() + "-" + bhkSizeList.get(bhkSizeList.size() - 1).intValue());
             }
         }
     }
